moving specifier classes to batfish-common-protocol; adding comments
