+++ projects/batfish/src/main/java/org/batfish/specifier/ConstantIpSpaceSpecifier.java
@@ -5,6 +5,10 @@ import com.google.common.collect.Multimap;
 import java.util.Set;
 import org.batfish.datamodel.IpSpace;
 
+/**
+ * An {@link IpSpaceSpecifier} that is constant (independent of the input {@link Location}). All
+ * input {@link Location}s are assigned the specified {@link IpSpace}.
+ */
 public class ConstantIpSpaceSpecifier implements IpSpaceSpecifier {
   private final IpSpace _ipSpace;
