--- projects/batfish/src/main/java/org/batfish/specifier/Location.java
@@ -1,5 +1,0 @@
-package org.batfish.specifier;
-
-public interface Location {
-  <T> T accept(LocationVisitor<T> visitor);
-}
