--- projects/batfish/src/main/java/org/batfish/specifier/LocationSpecifier.java
@@ -1,7 +1,0 @@
-package org.batfish.specifier;
-
-import java.util.Set;
-
-public interface LocationSpecifier {
-  Set<Location> resolve(SpecifierContext ctxt);
-}
