--- projects/batfish/src/main/java/org/batfish/specifier/LocationSpecifierFactory.java
@@ -1,7 +1,0 @@
-package org.batfish.specifier;
-
-public interface LocationSpecifierFactory {
-  String getName();
-
-  LocationSpecifier buildLocationSpecifier(Object input);
-}
